add vanity urls
