+++ TODO.md
@@ -1,5 +1,3 @@
 # TODO
 
-- Modify url params on text box change
-- Add vanity urls for three|four|etc-nines and nine-fives
 - /api and/or /json enpoint with simple json return
