+++ src/components/Input.svelte
@@ -13,6 +13,7 @@
     }
 
     borderstyle = "focus:ring-green-900 focus:border-green-900";
+    history.replaceState(null, '', `${window.location.origin}/${uptime}`)
     dispatch("update", uptime);
   }
 </script>
