+++ src/components/Math.svelte
@@ -1,19 +1,4 @@
 <script lang="typescript" context="module">
-  // get a number from the provided path or return default on failure
-  export function pathNumber(def: number): number {
-    const path = window.location.pathname;
-    if (path === "/") {
-      return def;
-    }
-
-    const pn = Number(path.replace(/^\/+/g, ""));
-    if (!Number.isNaN(pn) && pn >= 0 && pn <= 100) {
-      return pn;
-    }
-
-    return def;
-  }
-
   // allowed uptime in seconds given a percent and portion of the year
   export function uptimeSeconds(
     uptimePercent: number,
